ENH: ufunc dispatcher/concision
* use a ufunc workunit dispatcher to improve
concision of the ufunc code for these ufuncs,
as an initial demonstration:
  - `isfinite()` (`parallel_for`)
  - `equal()` (`parallel_for`)
  - `isinf()` (`parallel_for`)
  - `matmul()` (`parallel_reduce`)
  - `reciprocal()` (`parallel_for`)

* the reduction in LOC is around: 250

* the changes do not guarantee identical behavior
beyond the testsuite passing, so for example
error handling behavior may not be fully preserved

* our type system still needs a cleanup to make
things more consistent per gh-80
