+++ mypy.ini
@@ -104,3 +104,6 @@ ignore_missing_imports = True
 
 [mypy-pykokkos.lib.manipulate]
 ignore_errors = True
+
+[mypy-pykokkos.lib.ufunc_workunits]
+ignore_errors = True
